chore: enhancement sync proto script

+++ scripts/proto/sync_proto.sh
@@ -1,5 +1,22 @@
 #!/bin/bash
 
+# Init mode
+mode=remote
+DEPENDENCY_FILE="protodep.yaml"
+DEPENDENCY_LOCAL_FILE="protodep.local.yaml"
+
+while getopts 'm:f:' OPTION; do
+  case "$OPTION" in 
+    m)
+      mode=$OPTARG
+      ;;
+    f)
+      DEPENDENCY_FILE=$OPTARG
+      DEPENDENCY_LOCAL_FILE=$(echo $OPTARG | sed 's/yaml/local.yaml/')
+      ;;
+  esac
+done
+
 # Check if yq installed
 if ! command -v yq &> /dev/null
 then
@@ -18,27 +35,17 @@ then
     echo "✅ protobuf installed"
 fi
 
-# Get arguments
-for ARGUMENT in "$@"
-do
-   KEY=$(echo $ARGUMENT | cut -f1 -d=)
-
-   KEY_LENGTH=${#KEY}
-   VALUE="${ARGUMENT:$KEY_LENGTH+1}"
-
-   export "$KEY"="$VALUE"
-done
-
-# Check if DEPENDENCY_FILE is set
-if [ -z "$DEPENDENCY_FILE" ]
-then
-    DEPENDENCY_FILE="protodep.yaml"
-fi
-
 # Check if dependency file exists
 if [ ! -f $DEPENDENCY_FILE ]
 then
     echo "🙀 Dependency file not found."
+    echo "🙀 Exiting"
+    exit 1
+fi
+
+if [ ! -f $DEPENDENCY_LOCAL_FILE ] && [ $mode == "local" ]
+then
+    echo "🙀 Dependency local file not found."
     echo "🙀 Exiting"
     exit 1
 fi
@@ -49,67 +56,86 @@ if [ -f .env ]; then
 fi
 
 proto_yaml_file=$DEPENDENCY_FILE
+proto_yaml_local_file=$DEPENDENCY_LOCAL_FILE
 
 source=$(yq .source $proto_yaml_file)
 branch=$(yq .branch $proto_yaml_file)
 outpb=$(yq .outpb $proto_yaml_file)
 outdir=$(yq .outdir $proto_yaml_file)
 token_key=$(printenv $(yq .token_key $proto_yaml_file))
-dependencies=$(yq .dependencies $proto_yaml_file)
+remote_dependencies=$(yq .dependencies $proto_yaml_file)
+
+if [ $mode == "local" ]
+then
+  local_dependencies=$(yq .dependencies $proto_yaml_local_file)
+fi
 
 command=$(yq .command $proto_yaml_file)
 
-if [ -z "$source" ]
+if [[ -z "$source" || $source == null ]] && [ $mode == "remote" ]
 then
     echo "🙀 Source is not set. Please set outdir key of protodep.yaml"
     exit 2
 fi
 
-if [ -z "$branch" ]
+if [[ -z "$branch" || $branch == null ]] && [ $mode == "remote" ]
 then
     echo "🙀 Branch is not set. Please set branch key of protodep.yaml"
     exit 2
 fi
 
-if [ -z "$outpb" ]
+if [[ -z "$outpb" || $outpb == null ]] && [ $mode == "remote" ]
 then
     echo "🙀 OutPb is not set. Please set outpb key of protodep.yaml"
     exit 2
 fi
 
-if [ -z "$outdir" ]
+if [[ -z "$outdir" || $outdir == null ]] && [ $mode == "remote" ]
 then
     echo "🙀 Out dir is not set. Please set outdir key of protodep.yaml"
     exit 2
 fi
 
-if [ -z "$token_key" ]
+if [[ -z "$token_key" || $token_key == null ]] && [ $mode == "remote" ]
 then
     echo "🙀 Token key is not set. Please set token_key key of protodep.yaml"
     exit 2
 fi
 
-if [ -z "$dependencies" ]
+if [[ -z "$remote_dependencies" || $remote_dependencies == null ]] && [ $mode == "remote" ]
 then
     echo "🙀 Dependencies is not set. Please set dependencies key of protodep.yaml"
     exit 2
 fi
 
+if [[ -z "$local_dependencies"  || $local_dependencies == null ]] && [ $mode == "local" ]
+then
+    echo "🙀 Local Dependencies is not set. Please set dependencies key of protodep.local.yaml"
+    exit 2
+fi
+
 
 delete_out_dir() {
+  if [ $mode == "local" ];then
+    return
+  fi
+
   read -r -p "Are you sure to delete dir '$outpb' and '$outdir' ? [y|N] " response
   if [[ $response =~ (y|yes|Y) ]];then
     rm -rf $outpb/*.proto
     rm -rf $outdir/*
   else
-    echo "🙀 Exiting"
-    exit 3
+    return
   fi
 }
 
 delete_out_dir
 
 fetch_proto() {
+  if [ $mode == "local" ];then
+    return
+  fi
+
   dependency=$1
   name=$(parse_name $dependency)
 
@@ -121,6 +147,14 @@ fetch_proto() {
 
 parse_name() {
   echo $(echo $1 | cut -d "/" -f 2)
+}
+
+get_dependencies_list() {
+  if [ $mode == "local" ];then
+    echo "$(yq eval '.dependencies' $proto_yaml_local_file)"
+  else
+    echo "$(yq eval '.dependencies' $proto_yaml_file)"
+  fi
 }
 
 while IFS= read -r value; do
@@ -138,4 +172,4 @@ while IFS= read -r value; do
   )
 
   $($formatted_command)
-done < <(yq eval '.dependencies' $proto_yaml_file)
+done < <(get_dependencies_list)
